+++ src/helpers/aws-device-farm.js
@@ -69,7 +69,7 @@ class AWSDeviceFarm extends AWSBase {
     }
 
     //type : "ANDROID_APP" || "IOS_APP" || "WEB_APP" || "EXTERNAL_DATA" || "APPIUM_JAVA_JUNIT_TEST_PACKAGE" || "APPIUM_JAVA_TESTNG_TEST_PACKAGE" || "APPIUM_PYTHON_TEST_PACKAGE" || "APPIUM_NODE_TEST_PACKAGE" || "APPIUM_RUBY_TEST_PACKAGE" || "APPIUM_WEB_JAVA_JUNIT_TEST_PACKAGE" || "APPIUM_WEB_JAVA_TESTNG_TEST_PACKAGE" || "APPIUM_WEB_PYTHON_TEST_PACKAGE" || "APPIUM_WEB_NODE_TEST_PACKAGE" || "APPIUM_WEB_RUBY_TEST_PACKAGE" || "CALABASH_TEST_PACKAGE" || "INSTRUMENTATION_TEST_PACKAGE" || "UIAUTOMATION_TEST_PACKAGE" || "UIAUTOMATOR_TEST_PACKAGE" || "XCTEST_TEST_PACKAGE" || "XCTEST_UI_TEST_PACKAGE" || "APPIUM_JAVA_JUNIT_TEST_SPEC" || "APPIUM_JAVA_TESTNG_TEST_SPEC" || "APPIUM_PYTHON_TEST_SPEC" || "APPIUM_NODE_TEST_SPEC" || "APPIUM_RUBY_TEST_SPEC" || "APPIUM_WEB_JAVA_JUNIT_TEST_SPEC" || "APPIUM_WEB_JAVA_TESTNG_TEST_SPEC" || "APPIUM_WEB_PYTHON_TEST_SPEC" || "APPIUM_WEB_NODE_TEST_SPEC" || "APPIUM_WEB_RUBY_TEST_SPEC" || "INSTRUMENTATION_TEST_SPEC" || "XCTEST_UI_TEST_SPEC",
-    async listUploads(projectArn, type, category = 'PRIVATE') {
+    async listUploads(projectArn, type, status = 'SUCCEEDED', category = 'PRIVATE') {
         const params = {
             arn: projectArn,
             type: type
@@ -77,8 +77,8 @@ class AWSDeviceFarm extends AWSBase {
 
         const cmd = new ListUploadsCommand(params);
         let data = await this.applyCommand(cmd);
-        if (data && data.$metadata.httpStatusCode == 200) {
-            return data.uploads.filter((upload) => upload.category == category);
+        if (data && data.$metadata.httpStatusCode == 200 && data.uploads && data.uploads.length > 0) {
+            return data.uploads.filter((upload) => upload.category == category && upload.status == status);
         }
         return null;
     }
@@ -222,6 +222,11 @@ class AWSDeviceFarm extends AWSBase {
                 (uploadStatus.status == 'INITIALIZED' || uploadStatus.status == 'PROCESSING') &&
                 count < 10
             );
+
+            if (uploadStatus.status != 'SUCCEEDED') {
+                console.error(`Upload failed:  status "${uploadStatus.status}" - error "${uploadStatus.message}"`);
+                return null;
+            }
 
             const output = {
                 uploadArn: uploadOutput.arn,